Revert "GENAI-2318 Add Contextual Ranking support with GCS data artifacts (#1195)"
This reverts commit ff4a3166d875fb08481f973410286ba534532df4.
