+++ docs/data.md
@@ -185,12 +185,6 @@ The following additional metrics are recorded when curated recommendations are r
 - `recommendation.prior.last_updated` -
  A gauge for the staleness (in seconds) of the prior data, measured between when the data was
  updated in GCS and the current time.
-- `recommendation.ml.contextual.update.timing` - A timer to measure the duration (in ms) of updating the engagement data from GCS.
-- `recommendation.ml.contextual.size` - A gauge to track the size of the engagement blob on GCS.
-- `recommendation.ml.contextual.last_updated` -
- A gauge for the staleness (in seconds) of the contextual engagement data, measured between when the data was
- updated in GCS and the current time.
-
 
 ### Manifest
