+++ merino/configs/__init__.py
@@ -52,7 +52,6 @@ _validators = [
         "curated_recommendations.gcs.engagement.cron_interval_seconds",
         "curated_recommendations.gcs.prior.max_size",
         "curated_recommendations.gcs.prior.cron_interval_seconds",
-        "ml_recommendations.gcs.max_size",
         is_type_of=int,
         must_exist=True,
         env=["production", "staging", "development"],
@@ -62,14 +61,6 @@ _validators = [
         "curated_recommendations.gcs.gcp_project",
         "curated_recommendations.gcs.engagement.blob_name",
         "curated_recommendations.gcs.prior.blob_name",
-        is_type_of=str,
-        must_exist=True,
-        env=["production", "staging", "development"],
-    ),
-    Validator(
-        "ml_recommendations.gcs.bucket_name",
-        "ml_recommendations.gcs.gcp_project",
-        "ml_recommendations.gcs.blob_name",
         is_type_of=str,
         must_exist=True,
         env=["production", "staging", "development"],
