+++ merino/configs/production.toml
@@ -74,12 +74,3 @@ cache = "redis"
 # The store used for flightaware data. Either `redis` or `none`.
 # If `redis`, the global Redis settings must be set. See redis.server.
 cache = "redis"
-
-[production.ml_recommendations.gcs]
-# MERINO__ML_RECOMMENDATIONS__GCS__GCP_PROJECT
-# GCP project name where the GCS bucket lives.
-gcp_project = "moz-fx-merino-prod-5de4"
-
-# MERINO__ML_RECOMMENDATIONS__GCS__BUCKET_NAME
-# GCS bucket that contains ranked articles from ml
-bucket_name = "merino-ml-data-prod"
