+++ merino/configs/stage.toml
@@ -52,12 +52,3 @@ bucket_name = "merino-airflow-data-stagepy"
 # MERINO__CURATED_RECOMMENDATIONS__GCS__GCP_PROJECT
 # GCP project name where the GCS bucket lives.
 gcp_project = "moz-fx-merino-nonprod-ee93"
-
-[stage.ml_recommendations.gcs]
-# MERINO__ML_RECOMMENDATIONS__GCS__GCP_PROJECT
-# GCP project name where the GCS bucket lives.
-gcp_project = "moz-fx-merino-nonprod-db57"
-
-# MERINO__ML_RECOMMENDATIONS__GCS__BUCKET_NAME
-# GCS bucket that contains ranked articles from ml
-bucket_name = "merino-ml-data-stage"
