+++ merino/curated_recommendations/protocol.py
@@ -8,7 +8,6 @@ from datetime import datetime
 
 import numpy as np
 from pydantic import (
-    ConfigDict,
     Field,
     field_validator,
     model_validator,
@@ -16,7 +15,6 @@ from pydantic import (
     ValidationInfo,
     RootModel,
 )
-from pydantic.alias_generators import to_snake
 
 from merino.curated_recommendations.corpus_backends.protocol import (
     CorpusItem,
@@ -100,9 +98,6 @@ class ExperimentName(str, Enum):
     CONTEXTUAL_AD_V2_BETA_EXPERIMENT = "new-tab-contextual-ad-updates-v2-beta"
     CONTEXTUAL_AD_RELEASE_EXPERIMENT = "new-tab-contextual-ad-updates-release"
     CONTEXTUAL_AD_V2_RELEASE_EXPERIMENT = "new-tab-contextual-ad-updates-v2-release"
-    NEW_TAB_CUSTOM_SECTIONS_EXPERIMENT = "new-tab-custom-sections"
-    CONTEXTUAL_RANKING_CONTENT_EXPERIMENT = "content-contextual-ranking"
-
     # Experiment for doing local reranking of popular today via inferred interests
     INFERRED_LOCAL_EXPERIMENT = "new-tab-automated-personalization-local-ranking"
     INFERRED_LOCAL_EXPERIMENT_V2 = "new-tab-automated-personalization-local-ranking-2"
@@ -271,11 +266,6 @@ class CuratedRecommendation(CorpusItem):
 
 class CuratedRecommendationsRequest(BaseModel):
     """Body schema for requesting a list of curated recommendations"""
-
-    model_config = ConfigDict(
-        alias_generator=to_snake,
-        populate_by_name=True,
-    )
 
     locale: Locale
     region: str | None = None
