+++ merino/curated_recommendations/provider.py
@@ -3,7 +3,7 @@
 import logging
 from typing import cast
 
-from merino.curated_recommendations import LocalModelBackend, MLRecsBackend
+from merino.curated_recommendations import LocalModelBackend
 from merino.curated_recommendations.ml_backends.protocol import LOCAL_MODEL_MODEL_ID_KEY
 from merino.curated_recommendations.corpus_backends.protocol import (
     ScheduledSurfaceProtocol,
@@ -25,7 +25,7 @@ from merino.curated_recommendations.protocol import (
 from merino.curated_recommendations.rankers import (
     boost_preferred_topic,
     spread_publishers,
-    ThompsonSamplingRanker,
+    thompson_sampling,
 )
 from merino.curated_recommendations.sections import get_sections
 from merino.curated_recommendations.utils import (
@@ -52,14 +52,12 @@ class CuratedRecommendationsProvider:
         prior_backend: PriorBackend,
         sections_backend: SectionsProtocol,
         local_model_backend: LocalModelBackend,
-        ml_recommendations_backend: MLRecsBackend,
     ) -> None:
         self.scheduled_surface_backend = scheduled_surface_backend
         self.engagement_backend = engagement_backend
         self.prior_backend = prior_backend
         self.sections_backend = sections_backend
         self.local_model_backend = local_model_backend
-        self.ml_recommendations_backend = ml_recommendations_backend
 
     @staticmethod
     def is_sections_experiment(
@@ -88,11 +86,11 @@ class CuratedRecommendationsProvider:
         @param request: The full API request with all the data
         @return: A re-ranked list of curated recommendations
         """
-        ranker = ThompsonSamplingRanker(
-            engagement_backend=self.engagement_backend, prior_backend=self.prior_backend
-        )
-        recommendations = ranker.rank_items(
+        # 3. Apply Thompson sampling to rank recommendations by engagement
+        recommendations = thompson_sampling(
             recommendations,
+            engagement_backend=self.engagement_backend,
+            prior_backend=self.prior_backend,
             region=derive_region(request.locale, request.region),
         )
 
@@ -145,7 +143,6 @@ class CuratedRecommendationsProvider:
                 personal_interests=inferred_interests,
                 prior_backend=self.prior_backend,
                 sections_backend=self.sections_backend,
-                ml_backend=self.ml_recommendations_backend,
                 scheduled_surface_backend=self.scheduled_surface_backend,
                 region=derive_region(request.locale, request.region),
             )
