--- merino/curated_recommendations/rankers/__init__.py
@@ -1,23 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-"""Module dedicated to ranking algorithms for curated recommendations."""
-
-from merino.curated_recommendations.rankers.t_sampling import ThompsonSamplingRanker
-from merino.curated_recommendations.rankers.contextual_ranker import ContextualRanker
-from merino.curated_recommendations.rankers.ranker import Ranker
-from merino.curated_recommendations.rankers.utils import (
-    REGION_ENGAGEMENT_WEIGHT,
-    TOP_STORIES_SECTION_KEY,
-    boost_followed_sections,
-    filter_fresh_items_with_probability,
-    renumber_sections,
-    is_section_recently_followed,
-    renumber_recommendations,
-    spread_publishers,
-    boost_preferred_topic,
-    put_top_stories_first,
-    greedy_personalized_section_rank,
-    takedown_reported_recommendations,
-)
