+++ tests/integration/api/v1/curated_recommendations/test_curated_recommendations.py
@@ -27,7 +27,6 @@ from merino.curated_recommendations import (
     ConstantPrior,
     interest_picker,
     LocalModelBackend,
-    MLRecsBackend,
 )
 from merino.curated_recommendations.legacy.provider import LegacyCuratedRecommendationsProvider
 from merino.curated_recommendations.corpus_backends.protocol import (
@@ -41,13 +40,10 @@ from merino.curated_recommendations.engagement_backends.protocol import (
 )
 from merino.curated_recommendations.localization import LOCALIZED_SECTION_TITLES
 from merino.curated_recommendations.ml_backends.static_local_model import (
-    CONTEXTUAL_RANKING_TREATMENT_COUNTRY,
-    CONTEXTUAL_RANKING_TREATMENT_TZ,
     CTR_LIMITED_TOPIC_MODEL_ID_V1_B,
     DEFAULT_PRODUCTION_MODEL_ID,
 )
 from merino.curated_recommendations.ml_backends.protocol import (
-    ContextualArticleRankings,
     InferredLocalModel,
     ModelData,
     ModelType,
@@ -69,16 +65,6 @@ from merino.providers.manifest import get_provider as get_manifest_provider
 from merino.providers.manifest.backends.protocol import Domain
 from tests.types import FilterCaplogFixture
 
-# Music, NFL, Movies, Soccer, NBA
-REC_HIGH_CTR_IDS = [
-    "b2c10703-5377-4fe8-89d3-32fbd7288187",
-    "f509393b-c1d6-4500-8ed2-29f8a23f39a7",
-    "2afcef43-4663-446e-9d69-69cbc6966162",
-    "dc4b30c4-170b-4e9f-a068-bdc51474a0fb",
-    "9261e868-beff-4419-8071-7750d063d642",
-    "63909b8c-a619-45f3-9ebc-fd8fcaeb72b1",
-]
-
 
 def is_manual_section(section_id: str) -> bool:
     """Check if section ID is a UUID (manually created sections use UUIDs, ML sections use human-readable IDs).
@@ -90,24 +76,6 @@ def is_manual_section(section_id: str) -> bool:
         return True
     except ValueError:
         return False
-
-
-class MockMLRecommendationsBackend(MLRecsBackend):
-    """Mock class implementing the protocol for MLRecsBackend."""
-
-    def __init__(self):
-        super().__init__()
-
-    def get(
-        self, region: str | None = None, utcOffset: str | None = None
-    ) -> ContextualArticleRankings | None:
-        """Return sample ML recommendations"""
-        VERY_HIGH_CTR = {k: [1.0, 1.0] for k in REC_HIGH_CTR_IDS}
-
-        return ContextualArticleRankings(
-            granularity="",
-            shards={**VERY_HIGH_CTR, "41111154-ebb1-45d9-9799-a882f13cd8cc": [0.1, 0.1]},
-        )
 
 
 class MockEngagementBackend(EngagementBackend):
@@ -145,15 +113,29 @@ class MockEngagementBackend(EngagementBackend):
                 report_count=reports,
             )
 
-        VERY_HIGH_CTR = {
-            k: (
-                1_000_000 * self.experiment_traffic_fraction,
-                1_000_000 * self.experiment_traffic_fraction,
-            )
-            for k in REC_HIGH_CTR_IDS
-        }
         HIGH_CTR_ITEMS = {
-            **VERY_HIGH_CTR,
+            "b2c10703-5377-4fe8-89d3-32fbd7288187": (
+                1_000_000 * self.experiment_traffic_fraction,
+                1_000_000 * self.experiment_traffic_fraction,
+            ),  # ML music 100% CTR
+            "f509393b-c1d6-4500-8ed2-29f8a23f39a7": (
+                1_000_000 * self.experiment_traffic_fraction,
+                1_000_000 * self.experiment_traffic_fraction,
+            ),  # ML NFL 100% CTR
+            "2afcef43-4663-446e-9d69-69cbc6966162": (
+                1_000_000 * self.experiment_traffic_fraction,
+                1_000_000 * self.experiment_traffic_fraction,
+            ),  # ML Movies 100% CTR
+            "dc4b30c4-170b-4e9f-a068-bdc51474a0fb": (
+                1_000_000 * self.experiment_traffic_fraction,
+                1_000_000 * self.experiment_traffic_fraction,
+            ),  # ML Soccer 100% CTR
+            "9261e868-beff-4419-8071-7750d063d642": (
+                1_000_000 * self.experiment_traffic_fraction,
+                1_000_000 * self.experiment_traffic_fraction,
+            ),  # ML NBA 100% CTR
+            "63909b8c-a619-45f3-9ebc-fd8fcaeb72b1": (1_000_000, 1_000_000),  # ML Food 100% CTR
+            # The above 6 ML recs have the highest CTR & will be included in top_stories_section
             "41111154-ebb1-45d9-9799-a882f13cd8cc": (
                 990_000 * self.experiment_traffic_fraction,
                 1_000_000 * self.experiment_traffic_fraction,
@@ -253,12 +235,6 @@ def engagement_backend():
 
 
 @pytest.fixture
-def ml_recommendations_backend():
-    """Fixture for the MockMLRecommendationsBackend for standard use case"""
-    return MockMLRecommendationsBackend()
-
-
-@pytest.fixture
 def engagement_backend_legacy_sections_us():
     """Fixture for the MockEngagementBackend for an experiment that has a fraction of traffic"""
     return MockEngagementBackend(SECTIONS_HOLDBACK_TOTAL_PERCENT)
@@ -292,7 +268,6 @@ def provider(
     engagement_backend: EngagementBackend,
     prior_backend: PriorBackend,
     local_model_backend: LocalModelBackend,
-    ml_recommendations_backend: MLRecsBackend,
 ) -> CuratedRecommendationsProvider:
     """Mock curated recommendations provider."""
     return CuratedRecommendationsProvider(
@@ -301,7 +276,6 @@ def provider(
         prior_backend=prior_backend,
         sections_backend=sections_backend,
         local_model_backend=local_model_backend,
-        ml_recommendations_backend=ml_recommendations_backend,
     )
 
 
@@ -572,28 +546,23 @@ class TestCuratedRecommendationsRequestParameters:
         )
         assert response.status_code == 400
 
-    @pytest.mark.parametrize("utcOffset", [0, 12, 24])
-    def test_curated_recommendations_valid_utc_offset(self, utcOffset, client: TestClient):
-        """Test the curated recommendations endpoint accepts valid utc_offset values."""
-        response = client.post(
-            "/api/v1/curated-recommendations",
-            json={"locale": Locale.EN_US, "utc_offset": utcOffset},
-        )
-        assert response.status_code == 200
-
-    @pytest.mark.parametrize("utcOffset", [-1, 11.5, 25, "Z"])
-    def test_curated_recommendations_invalid_utc_offset(self, utcOffset, client: TestClient):
-        """Test the curated recommendations endpoint rejects invalid utc_offset values
-        and is looking both camel and snake inputs.
+    @pytest.mark.parametrize("utc_offset", [0, 12, 24])
+    def test_curated_recommendations_valid_utc_offset(self, utc_offset, client: TestClient):
+        """Test the curated recommendations endpoint accepts valid utc_offset values.
+        This includes values that require rounding (e.g., 3.7 should be rounded to 4).
         """
         response = client.post(
             "/api/v1/curated-recommendations",
-            json={"locale": Locale.EN_US, "utc_offset": utcOffset},
+            json={"locale": Locale.EN_US, "utcOffset": utc_offset},
         )
-        assert response.status_code == 400
+        assert response.status_code == 200
+
+    @pytest.mark.parametrize("utc_offset", [-1, 11.5, 25, "Z"])
+    def test_curated_recommendations_invalid_utc_offset(self, utc_offset, client: TestClient):
+        """Test the curated recommendations endpoint rejects invalid utc_offset values."""
         response = client.post(
             "/api/v1/curated-recommendations",
-            json={"locale": Locale.EN_US, "utcOffset": utcOffset},
+            json={"locale": Locale.EN_US, "utcOffset": utc_offset},
         )
         assert response.status_code == 400
 
@@ -1517,52 +1486,6 @@ class TestSections:
                 assert sections[tech_stuff_id]["title"] == "Tech stuff"
 
     @pytest.mark.parametrize(
-        "experiment_branch",
-        [
-            CONTEXTUAL_RANKING_TREATMENT_TZ,
-            CONTEXTUAL_RANKING_TREATMENT_COUNTRY,
-        ],
-    )
-    def test_sections_contextual_ranking(self, client: TestClient, experiment_branch):
-        """Test that sections feed includes both manually created and ML-generated sections for contextual ranking.
-
-        Both MANUAL and ML sections should be returned together.
-        """
-        response = client.post(
-            "/api/v1/curated-recommendations",
-            json={
-                "locale": "en-US",
-                "feeds": ["sections"],
-                "experimentName": ExperimentName.CONTEXTUAL_RANKING_CONTENT_EXPERIMENT.value,
-                "experimentBranch": experiment_branch,
-            },
-        )
-        data = response.json()
-
-        # Check if the response is valid
-        assert response.status_code == 200
-
-        feeds = data["feeds"]
-        sections = {name: section for name, section in feeds.items() if section is not None}
-
-        # top_stories_section should always be present
-        assert "top_stories_section" in sections
-
-        # Should have ML sections (legacy topics)
-        legacy_topics = {topic.value for topic in Topic}
-        ml_sections_found = [sid for sid in sections if sid in legacy_topics]
-        assert len(ml_sections_found) > 0, "Should have at least some ML legacy topic sections"
-
-        # Check if any manually created sections appear (they may or may not, depending on
-        # whether they have enough items after top stories are removed)
-        manual_sections = [sid for sid in sections if is_manual_section(sid)]
-        if manual_sections:
-            # If the "Tech stuff" manual section appears, verify it has the correct title
-            tech_stuff_id = "d532b687-108a-4edb-a076-58a6945de714"
-            if tech_stuff_id in sections:
-                assert sections[tech_stuff_id]["title"] == "Tech stuff"
-
-    @pytest.mark.parametrize(
         "sections_payload",
         [
             {},
@@ -2106,7 +2029,7 @@ class TestSections:
                 "experimentName": "optin-new-tab-ml-sections",
                 "experimentBranch": "treatment",
                 "utc_offset": 17,
-                "coarse_os": "win",
+                "coarse_os": "windows",
                 "surface_id": "",
                 "locale": "en-US",
                 "region": "US",
@@ -2115,6 +2038,7 @@ class TestSections:
             },
         )
         data = response.json()
+
         interest_picker_response = data["interestPicker"]
         if enable_interest_picker:
             assert interest_picker_response is not None
