+++ tests/unit/curated_recommendations/test_rankers.py
@@ -30,10 +30,6 @@ from merino.curated_recommendations.prior_backends.engagment_rescaler import (
 )
 from merino.curated_recommendations.prior_backends.protocol import Prior, PriorBackend
 from merino.curated_recommendations.rankers import (
-    filter_fresh_items_with_probability,
-    ThompsonSamplingRanker,
-)
-from merino.curated_recommendations.rankers.utils import (
     spread_publishers,
     boost_preferred_topic,
     boost_followed_sections,
@@ -42,6 +38,8 @@ from merino.curated_recommendations.rankers.utils import (
     put_top_stories_first,
     greedy_personalized_section_rank,
     takedown_reported_recommendations,
+    thompson_sampling,
+    filter_fresh_items_with_probability,
 )
 from tests.unit.curated_recommendations.fixtures import (
     generate_recommendations,
@@ -266,7 +264,7 @@ class TestFilterFreshItemsWithProbability:
             except StopIteration:
                 return 0.0
 
-        monkeypatch.setattr("merino.curated_recommendations.rankers.utils.random", fake_random)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.random", fake_random)
 
         filtered, backlog = filter_fresh_items_with_probability(
             recs, fresh_story_prob=0.5, max_items=3
@@ -293,7 +291,7 @@ class TestFilterFreshItemsWithProbability:
             except StopIteration:
                 return 0.9
 
-        monkeypatch.setattr("merino.curated_recommendations.rankers.utils.random", fake_random)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.random", fake_random)
 
         filtered, backlog = filter_fresh_items_with_probability(
             recs, fresh_story_prob=0.1, max_items=1
@@ -354,11 +352,14 @@ class TestThompsonSampling:
         rescaler = CrawledContentRescaler()
 
         # Make beta sampling deterministic to avoid flakiness.
-        monkeypatch.setattr(
-            "merino.curated_recommendations.rankers.t_sampling.beta.rvs", lambda a, b: 0.42
+        monkeypatch.setattr("merino.curated_recommendations.rankers.beta.rvs", lambda a, b: 0.42)
+
+        ranked = thompson_sampling(
+            recs,
+            engagement_backend=engagement_backend,
+            prior_backend=prior_backend,
+            rescaler=rescaler,
         )
-        ranker = ThompsonSamplingRanker(engagement_backend, prior_backend)
-        ranked = ranker.rank_items(recs, rescaler)
 
         assert len(ranked) == 2
         by_id = {rec.corpusItemId: rec for rec in ranked}
@@ -393,12 +394,12 @@ class TestThompsonSampling:
         rescaler.fresh_items_top_stories_max_percentage = 0
 
         # Make beta sampling deterministic to avoid flakiness.
-        monkeypatch.setattr(
-            "merino.curated_recommendations.rankers.t_sampling.beta.rvs", lambda a, b: 0.42
-        )
-        ranker = ThompsonSamplingRanker(engagement_backend, prior_backend)
-        ranked = ranker.rank_items(
+        monkeypatch.setattr("merino.curated_recommendations.rankers.beta.rvs", lambda a, b: 0.42)
+
+        ranked = thompson_sampling(
             recs,
+            engagement_backend=engagement_backend,
+            prior_backend=prior_backend,
             rescaler=rescaler,
         )
