+++ tests/unit/curated_recommendations/test_sections.py
@@ -22,7 +22,6 @@ from merino.curated_recommendations.layouts import (
     layout_4_medium,
     layout_6_tiles,
 )
-from merino.curated_recommendations.prior_backends.constant_prior import ConstantPrior
 from merino.curated_recommendations.prior_backends.engagment_rescaler import (
     SchedulerHoldbackRescaler,
     CrawledContentRescaler,
@@ -34,7 +33,7 @@ from merino.curated_recommendations.protocol import (
     CuratedRecommendation,
     RankingData,
 )
-from merino.curated_recommendations.rankers import ThompsonSamplingRanker
+from merino.curated_recommendations.rankers import section_thompson_sampling
 from merino.curated_recommendations.sections import (
     adjust_ads_in_sections,
     exclude_recommendations_from_blocked_sections,
@@ -703,7 +702,7 @@ class TestGetTopStoryList:
             rec.ranking_data = RankingData(alpha=1, beta=1, score=1, is_fresh=True)
         rescaler = CrawledContentRescaler(fresh_items_top_stories_max_percentage=0.5)
 
-        monkeypatch.setattr("merino.curated_recommendations.rankers.utils.random", lambda: 0.1)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.random", lambda: 0.1)
 
         result = get_top_story_list(
             items, top_count=3, extra_count=0, extra_source_depth=0, rescaler=rescaler
@@ -722,7 +721,7 @@ class TestGetTopStoryList:
             rec.ranking_data = RankingData(alpha=1, beta=1, score=1, is_fresh=True)
         rescaler = CrawledContentRescaler(fresh_items_top_stories_max_percentage=0.5)
 
-        monkeypatch.setattr("merino.curated_recommendations.rankers.utils.random", lambda: 0.9)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.random", lambda: 0.9)
 
         result = get_top_story_list(
             items, top_count=2, extra_count=0, extra_source_depth=0, rescaler=rescaler
@@ -801,18 +800,18 @@ class TestSectionThompsonSampling:
             )
         }
 
-        monkeypatch.setattr(
-            "merino.curated_recommendations.rankers.t_sampling.beta.rvs", lambda a, b: 0.5
-        )
-        monkeypatch.setattr("merino.curated_recommendations.rankers.utils.random", lambda: 0.8)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.beta.rvs", lambda a, b: 0.5)
+        monkeypatch.setattr("merino.curated_recommendations.rankers.random", lambda: 0.8)
 
-        top_n = 4
+        section_thompson_sampling(sections, backend, top_n=4, rescaler=rescaler)
 
-        thomspon_sampling = ThompsonSamplingRanker(backend, ConstantPrior())
-        thomspon_sampling.rank_sections(sections, top_n=top_n, rescaler=rescaler)
-
-        # Engagement tracker checks for global and regional engagement
-        assert len(backend.requests) == top_n * 2
+        expected_ids = [
+            recs[3].corpusItemId,
+            recs[4].corpusItemId,
+            recs[0].corpusItemId,
+            recs[1].corpusItemId,
+        ]
+        assert backend.requests == expected_ids
 
 
 class TestCycleLayoutsForRankedSections:
